feat: refactoring state to redux actions and resolving remenant point
