+++ frontend/src/features/InterestPoint/useCases/saveInterestPointFeature.ts
@@ -3,6 +3,8 @@ import { getGeoJSONFromFeature } from 'domain/types/map'
 import Feature from 'ol/Feature'
 import Point from 'ol/geom/Point'
 
+import type { LineString } from 'ol/geom'
+
 export const saveInterestPointFeature = (feature?: Feature | undefined) => (dispatch, getState) => {
   const { currentInterestPoint }: InterestPointState = getState().interestPoint
 
@@ -19,7 +21,8 @@ export const saveInterestPointFeature = (feature?: Feature | undefined) => (disp
 
   featureToSave.setId(currentInterestPoint?.uuid)
 
-  const geoJSONFeature = getGeoJSONFromFeature(featureToSave)
+  // FIXME: had to force cast
+  const geoJSONFeature = getGeoJSONFromFeature(featureToSave) as unknown as Feature<LineString>
 
   const { feature: currentFeature, ...currentInterestPointWithoutFeature } = currentInterestPoint
