+++ frontend/src/features/InterestPoint/useCases/updateName.ts
@@ -1,0 +1,13 @@
+import { updateCurrentInterestPoint, type InterestPointState } from '../slice'
+
+export const updateNameAction = (name: string | undefined) => (dispatch, getState) => {
+  const { currentInterestPoint }: InterestPointState = getState().interestPoint
+
+  if (currentInterestPoint?.name !== name) {
+    const updatedName = name === undefined ? null : name
+
+    const { name: currentName, ...currentInterestPointWithoutName } = currentInterestPoint
+
+    dispatch(updateCurrentInterestPoint({ name: updatedName, ...currentInterestPointWithoutName }))
+  }
+}
